requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-jose==3.3.0
pytest
jwt==1.3.1
python-dotenv
stripe
